Patch from MoDaX which fixes a compile error on arm architecture
svn path=/trunk/extragear/network/ktorrent/; revision=841546

+++ ktorrent/recommendedsettingsdlg.cpp
@@ -155,7 +155,7 @@ namespace kt
 			max_downloads = 1;
 		
 		max_conn_glob = qRound(qMin((double)pow((int)(upload_rate*8),0.8)+50,900.0));
-		max_conn_tor = qRound(qMin(max_conn_glob * 1.2 / max_torrents,(qreal)max_conn_glob));
+		max_conn_tor = qRound(qMin((qreal)(max_conn_glob * 1.2 / max_torrents),(qreal)max_conn_glob));
 	
 		m_max_upload->setText(QString("<b>%1</b>").arg(KBytesPerSecToString(max_upload_speed)));
 		m_max_download->setText(QString("<b>%1</b>").arg(KBytesPerSecToString(max_download_speed)));
